fix number formatting, add duration into timer state
