+++ components/bsb/timer/active/impl/src/commonMain/kotlin/com/flipperdevices/bsb/timer/active/composable/TimerActiveComposableScreen.kt
@@ -1,125 +1,146 @@
 package com.flipperdevices.bsb.timer.active.composable
 
+
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.flipperdevices.bsb.core.theme.BusyBarThemeInternal
+import com.flipperdevices.bsb.core.theme.LocalBusyBarFonts
 import com.flipperdevices.bsb.core.theme.LocalPallet
-import com.flipperdevices.bsb.timer.background.model.ControlledTimerState
 import com.flipperdevices.bsb.timer.common.composable.appbar.ButtonTimerComposable
 import com.flipperdevices.bsb.timer.common.composable.appbar.ButtonTimerState
+import com.flipperdevices.bsb.timer.common.composable.appbar.StatusLowBarComposable
 import com.flipperdevices.bsb.timer.common.composable.appbar.StatusType
 import com.flipperdevices.bsb.timer.common.composable.appbar.TimerAppBarComposable
-import com.flipperdevices.core.data.timer.TimerState
 import com.flipperdevices.ui.button.BChipButton
-import org.jetbrains.compose.ui.tooling.preview.Preview
+import com.flipperdevices.ui.timeline.toFormattedTime
+import kotlin.time.Duration
 import kotlin.time.Duration.Companion.minutes
+import kotlin.time.Duration.Companion.seconds
+import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
 @Suppress("LongMethod")
-fun TimerActiveComposableScreen(
-    timerState: ControlledTimerState,
-    onPauseClick: () -> Unit,
-    onStopClick: () -> Unit,
-    onSkipClick: (() -> Unit)?,
+fun TimerOnComposableScreen(
+    timeLeft: Duration,
     modifier: Modifier = Modifier,
-    workPhaseText: String? = null
+    workPhaseText: String? = null,
+    onSkip: (() -> Unit)? = null,
+    onPauseClick: () -> Unit,
+    onBack: () -> Unit,
 ) {
     Box(
-        modifier = modifier
+        modifier = Modifier
             .fillMaxSize()
-            .background(Color(color = 0xFF191919)) // todo
-            .statusBarsPadding()
-            .navigationBarsPadding()
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(
-                    Brush.verticalGradient(
-                        colors = listOf(
-                            Color(color = 0xFFEF2828).copy(alpha = 0.1f), // todo
-                            Color.Transparent,
-                            Color.Transparent
-                        )
+            .background(
+                Brush.verticalGradient(
+                    colors = listOf(
+                        Color(color = 0xFF900606), // todo
+                        Color(color = 0xFF430303),
                     )
                 )
-        )
-        TimerAppBarComposable(
-            statusType = StatusType.BUSY,
-            modifier = Modifier.align(Alignment.TopCenter),
-            workPhaseText = workPhaseText
-        )
-
+            )
+    )
+    Box(modifier = modifier.fillMaxSize().navigationBarsPadding().statusBarsPadding()) {
         Column(
-            modifier = Modifier.align(Alignment.Center),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
+            modifier = Modifier.fillMaxWidth(),
+            verticalArrangement = Arrangement.spacedBy(36.dp),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            TimerTimeComposable(
-                timerState = timerState.timerState
+            TimerAppBarComposable(onBackClick = onBack)
+            StatusLowBarComposable(
+                type = StatusType.BUSY,
+                statusDesc = workPhaseText
             )
         }
 
         Column(
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier.fillMaxWidth().align(Alignment.BottomCenter)
+            modifier = Modifier.align(Alignment.Center),
+            verticalArrangement = Arrangement.spacedBy(16.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
         ) {
             Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 24.dp),
-                horizontalArrangement = Arrangement.spacedBy(12.dp),
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.Center,
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                val contentPadding = PaddingValues(
-                    vertical = 24.dp
-                )
-                ButtonTimerComposable(
-                    state = when {
-                        timerState.isOnPause -> ButtonTimerState.START
-                        else -> ButtonTimerState.PAUSE
-                    },
-                    onClick = onPauseClick,
-                    modifier = Modifier.weight(1f),
-                    contentPadding = contentPadding
-                )
-                ButtonTimerComposable(
-                    state = ButtonTimerState.STOP,
-                    onClick = onStopClick,
-                    modifier = Modifier.weight(1f),
-                    contentPadding = contentPadding
-                )
+                timeLeft.toComponents { days, hours, minutes, seconds, nanoseconds ->
+                    Text(
+                        text = "${minutes.toFormattedTime()}",
+                        style = TextStyle(
+                            fontSize = 64.sp,
+                            fontWeight = FontWeight.W500,
+                            fontFamily = LocalBusyBarFonts.current.jetbrainsMono,
+                            color = LocalPallet.current.white.invert
+                        )
+                    )
+                    Text(
+                        text = ":",
+                        style = TextStyle(
+                            fontSize = 64.sp,
+                            fontWeight = FontWeight.W500,
+                            fontFamily = LocalBusyBarFonts.current.jetbrainsMono,
+                            color = LocalPallet.current.white.invert
+                        )
+                    )
+                    Text(
+                        text = "${seconds.toFormattedTime()}",
+                        style = TextStyle(
+                            fontSize = 64.sp,
+                            fontWeight = FontWeight.W500,
+                            fontFamily = LocalBusyBarFonts.current.jetbrainsMono,
+                            color = LocalPallet.current.white.invert
+                        )
+                    )
+                }
             }
-            onSkipClick?.let {
+            onSkip?.let {
                 BChipButton(
-                    modifier = Modifier.fillMaxWidth(),
+                    onClick = onSkip,
+                    background = Color.Transparent,
                     text = "Skip",
                     painter = null,
-                    background = Color.Transparent,
-                    onClick = onSkipClick,
+                    fontSize = 17.sp,
                     contentColor = LocalPallet.current
                         .transparent
                         .whiteInvert
-                        .primary
-                        .copy(alpha = 0.5f)
+                        .secondary,
+                    modifier = Modifier.fillMaxWidth(0.4f)
                 )
             }
+        }
+        Column(
+            modifier = Modifier
+                .align(Alignment.BottomCenter)
+                .padding(24.dp),
+            verticalArrangement = Arrangement.spacedBy(12.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            ButtonTimerComposable(
+                modifier = Modifier.fillMaxWidth(0.6f),
+                state = ButtonTimerState.PAUSE,
+                onClick = onPauseClick
+            )
+            Spacer(Modifier.height(8.dp))
         }
     }
 }
@@ -128,16 +149,13 @@ fun TimerActiveComposableScreen(
 @Preview
 private fun MainScreenComposableScreenPreview() {
     BusyBarThemeInternal {
-        TimerActiveComposableScreen(
+        TimerOnComposableScreen(
             modifier = Modifier.fillMaxSize(),
-            timerState = ControlledTimerState(
-                timerState = TimerState(32.minutes),
-                isOnPause = false
-            ),
-            workPhaseText = "1/4\nwork phase",
+            workPhaseText = "1/4",
+            timeLeft = 13.minutes.plus(10.seconds),
+            onSkip = {},
             onPauseClick = {},
-            onSkipClick = {},
-            onStopClick = {},
+            onBack = {}
         )
     }
 }
