+++ components/bsb/timer/rest/impl/build.gradle
@@ -14,6 +14,7 @@ commonDependencies {
     implementation(projects.components.bsb.core.theme)
     implementation(projects.components.core.ui.sheet)
     implementation(projects.components.core.ui.button)
+    implementation(projects.components.core.ui.timeline)
 
     implementation(projects.components.bsb.preference.api)
     implementation(projects.components.bsb.timer.common)
