+++ components/bsb/timer/rest/impl/src/commonMain/kotlin/com/flipperdevices/bsb/timer/finish/api/FinishTimerScreenDecomposeComponentImpl.kt
@@ -28,6 +28,7 @@ class FinishTimerScreenDecomposeComponentImpl(
             onSkip = {},
             timeLeft = 2.minutes,
             statusType = StatusType.LONG_REST,
+            onBackClick = {}
         )
     }
